+++ lib/dependency_graph.h
@@ -30,8 +30,8 @@ namespace design {
         public:
             DependencyGraph(std::vector<std::string> structures, std::string constraints, R rand);
 
-            unsigned long long number_of_sequences();
-            unsigned long long number_of_sequences(int connected_component_ID);
+            double number_of_sequences();
+            double number_of_sequences(int connected_component_ID);
             bool is_bipartite() {
                 return bipartite;
             }
@@ -41,10 +41,10 @@ namespace design {
             void set_sequence_string(std::string seq_str);
             void set_sequence();
             // call this function to mutate a random subgraph (either a path, if graph_type=-1 or a connected component, if graph_type=1)
-            unsigned long long mutate_local_global(int graph_type, int min_num_pos, int max_num_pos);
-            unsigned long long mutate_global(int connected_component_ID);
-            unsigned long long mutate(int position);
-            unsigned long long mutate(int start, int end);
+            double mutate_local_global(int graph_type, int min_num_pos, int max_num_pos);
+            double mutate_global(int connected_component_ID);
+            double mutate(int position);
+            double mutate(int start, int end);
             std::map< int, std::vector<int> > connected_components();
             std::vector< int > special_vertices();
             R * rand_ptr;
@@ -54,9 +54,9 @@ namespace design {
             bool bipartite; // if dependency graph is bipartite and a therefore a solution exists
             R rand;
             void calculate_probabilities(Graph& g);
-            unsigned long long sample_sequence(Graph& g);
+            double sample_sequence(Graph& g);
             void reset_colors(Graph& g);
-            unsigned long long mutate(Graph& g);
+            double mutate(Graph& g);
             // this function fills the subgraphs set with all the sugraphs of the given type (root, cc, bc, path)
             // if int type= -1, then it returns all subgraphs which are actual paths (gp.is_path == true).
             // you can specify also a minimal and maximal size of the subgraph
