+++ tests/test_probability_matrix.cc
@@ -100,7 +100,7 @@ BOOST_AUTO_TEST_CASE(PermuteKey2) {
 BOOST_AUTO_TEST_CASE(GetPutKey1) {
 
     ProbabilityMatrix pm;
-    unsigned long long control = 0;
+    double control = 0;
 
     for (auto b : base_conversion[N]) {
         for (auto c : base_conversion[S]) {
@@ -128,7 +128,7 @@ BOOST_AUTO_TEST_CASE(GetPutKey1) {
 BOOST_AUTO_TEST_CASE(GetPutKey2) {
 
     ProbabilityMatrix pm;
-    unsigned long long control = 0;
+    double control = 0;
 
     for (auto b : base_conversion[N]) {
         for (auto c : base_conversion[S]) {
@@ -156,7 +156,7 @@ BOOST_AUTO_TEST_CASE(GetPutKey2) {
 BOOST_AUTO_TEST_CASE(GetPutKey3) {
 
     ProbabilityMatrix pm;
-    unsigned long long control = 0;
+    double control = 0;
 
     for (auto b : base_conversion[N]) {
         for (auto c : base_conversion[S]) {
@@ -184,7 +184,7 @@ BOOST_AUTO_TEST_CASE(GetPutKey3) {
 BOOST_AUTO_TEST_CASE(GetNOS1) {
 
     ProbabilityMatrix pm;
-    unsigned long long control = 0;
+    double control = 0;
 
     for (auto b : base_conversion[N]) {
         for (auto c : base_conversion[S]) {
@@ -206,7 +206,7 @@ BOOST_AUTO_TEST_CASE(GetNOS1) {
 BOOST_AUTO_TEST_CASE(GetNOS2) {
 
     ProbabilityMatrix pm;
-    unsigned long long control = 0;
+    double control = 0;
 
     BOOST_TEST_MESSAGE("Try to get number of sequences for empty ProbabilityMatrix");
 
@@ -269,7 +269,7 @@ BOOST_AUTO_TEST_CASE(MultiplyPM1) {
     z = x * y;
 
     ProbabilityKey zkey;
-    unsigned long long mnos = 0;
+    double mnos = 0;
     for (auto b : base_conversion[N]) {
         for (auto c : base_conversion[N]) {
             zkey[1] = b;
@@ -309,7 +309,7 @@ BOOST_AUTO_TEST_CASE(MultiplyPM2) {
     z = x * y;
 
     ProbabilityKey zkey;
-    unsigned long long mnos = 0;
+    double mnos = 0;
     for (auto b : base_conversion[N]) {
         for (auto c : base_conversion[N]) {
             zkey[1] = b;
@@ -349,7 +349,7 @@ BOOST_AUTO_TEST_CASE(MultiplyPM3) {
     z = x * y;
 
     ProbabilityKey zkey;
-    unsigned long long mnos = 0;
+    double mnos = 0;
     for (auto b : base_conversion[N]) {
         zkey[1] = b;
         BOOST_CHECK(z[zkey] == ((3 * b)*(5 + b)));
@@ -383,7 +383,7 @@ BOOST_AUTO_TEST_CASE(MultiplyPM4) {
     z = x * y;
 
     ProbabilityKey zkey;
-    unsigned long long mnos = 0;
+    double mnos = 0;
     for (auto b : base_conversion[N]) {
         zkey[1] = b;
         BOOST_CHECK(z[zkey] == 16 * (3 * b));
@@ -414,7 +414,7 @@ BOOST_AUTO_TEST_CASE(MultiplyPM5) {
     z = x * y;
 
     ProbabilityKey zkey;
-    unsigned long long mnos = 0;
+    double mnos = 0;
     BOOST_CHECK(z[zkey] == 12 * 24);
     mnos += z[zkey];
 
@@ -474,9 +474,9 @@ BOOST_AUTO_TEST_CASE(MakeInternal1) {
     ProbabilityKey testkey;
     testkey[1] = U;
     testkey[7] = C;
-    unsigned long long ytest = y[testkey];
+    double ytest = y[testkey];
     testkey[5] = N;
-    unsigned long long xtest = x[testkey];
+    double xtest = x[testkey];
     BOOST_CHECK(ytest == xtest);
     BOOST_CHECK(x.getSpecials().size() == y.getSpecials().size() + 1);
     BOOST_CHECK(x.mnos() == y.mnos());
@@ -501,9 +501,9 @@ BOOST_AUTO_TEST_CASE(MakeInternal2) {
     std::cerr << "y:" << std::endl << y << std::endl;
 
     ProbabilityKey testkey;
-    unsigned long long ytest = y[testkey];
+    double ytest = y[testkey];
     testkey[5] = N;
-    unsigned long long xtest = x[testkey];
+    double xtest = x[testkey];
     BOOST_CHECK(ytest == xtest);
     BOOST_CHECK(x.getSpecials().size() == y.getSpecials().size() + 1);
     BOOST_CHECK(x.mnos() == y.mnos());
@@ -533,7 +533,7 @@ BOOST_AUTO_TEST_CASE(RandomlySampleKey1) {
     constraint[4] = V;
     constraint[7] = N;
 
-    std::pair<ProbabilityKey, unsigned long long> result = m.sample(constraint, &rand_gen);
+    std::pair<ProbabilityKey, double> result = m.sample(constraint, &rand_gen);
     BOOST_CHECK(result.first[1] == A);
     BOOST_CHECK(result.first[4] == A);
     BOOST_CHECK(result.first[7] == U);
@@ -560,14 +560,14 @@ BOOST_AUTO_TEST_CASE(RandomlySampleKey2) {
     }
     ProbabilityMatrix c = m;
 
-    std::pair<ProbabilityKey, unsigned long long> result = m.sample(&rand_gen);
+    std::pair<ProbabilityKey, double> result = m.sample(&rand_gen);
     BOOST_CHECK(result.first[1] == C);
     BOOST_CHECK(result.first[4] == U);
     BOOST_CHECK(result.first[7] == U);
     BOOST_CHECK(result.second == 6191);
 
     for (int i = 0; i < 1000; i++) {
-        std::pair<ProbabilityKey, unsigned long long> test = m.sample(&rand_gen);
+        std::pair<ProbabilityKey, double> test = m.sample(&rand_gen);
         //std::cerr << "One hundred Test:" << std::endl << test << std::endl;
     }
     BOOST_CHECK(m.mnos() == c.mnos());
